refactor(workspace): own blueprint session materialization
